Make DictWrapper inherit from dict
As of this commit, DictWrapper is just a dict with `gets` and `getn`
added. Really only `getn` is useful right now, though leaving the class
abstraction as it could hold other useful metadata later on.

An alternative option for users is to inherit the functions directly and
just call those (i.e. `get(msg, key)`)