Handle None case for `_nested_get` unwrap
This handles the valid case where a user is trying to get optional data
via the `get` wrapper.

Will leave `_nested_delete` alone since that's only used internally
(currently) and the expected value should always be there (currently).
Based on that it makes sense to error loudly instead of silently, though
if the functionality gets publicly exposed then it'd make sense to have
it mimic the behavior in this commit